+++ packages/scaffold/src/views/w3m-connect-view/index.tsx
@@ -7,11 +7,22 @@ import {
 } from '@web3modal/core-react-native';
 import type { WcWallet } from '@web3modal/core-react-native';
 import { ListWallet, FlexView } from '@web3modal/ui-react-native';
+import { UiUtil } from '../../utils/UiUtil';
 
 export function ConnectView() {
-  const { recommended, featured, count } = useSnapshot(ApiController.state);
+  const { recommended, featured, installed, count } = useSnapshot(ApiController.state);
   const { recentWallets } = useSnapshot(ConnectionController.state);
   const imageHeaders = ApiController._getApiHeaders();
+
+  const RECENT_COUNT = installed.length ? 1 : 2;
+  const INSTALLED_COUNT =
+    installed.length >= UiUtil.TOTAL_VISIBLE_WALLETS
+      ? UiUtil.TOTAL_VISIBLE_WALLETS - RECENT_COUNT
+      : installed.length;
+
+  const FEATURED_COUNT = UiUtil.TOTAL_VISIBLE_WALLETS - RECENT_COUNT - INSTALLED_COUNT;
+  const RECOMMENDED_COUNT =
+    UiUtil.TOTAL_VISIBLE_WALLETS - RECENT_COUNT - INSTALLED_COUNT - FEATURED_COUNT;
 
   const onWalletPress = (wallet: WcWallet) => {
     RouterController.push('ConnectingWalletConnect', { wallet });
@@ -22,28 +33,52 @@ export function ConnectView() {
       return null;
     }
 
-    return recentWallets.map(wallet => (
-      <ListWallet
-        key={wallet?.id}
-        imageSrc={AssetUtil.getWalletImage(wallet)}
-        imageHeaders={imageHeaders}
-        name={wallet?.name ?? 'Unknown'}
-        onPress={() => onWalletPress(wallet!)}
-        tagLabel="Recent"
-        tagVariant="shade"
-      />
-    ));
+    return recentWallets
+      .slice(0, RECENT_COUNT)
+      .map(wallet => (
+        <ListWallet
+          key={wallet?.id}
+          imageSrc={AssetUtil.getWalletImage(wallet)}
+          imageHeaders={imageHeaders}
+          name={wallet?.name ?? 'Unknown'}
+          onPress={() => onWalletPress(wallet!)}
+          tagLabel="Recent"
+          tagVariant="shade"
+        />
+      ));
+  };
+
+  const installedTemplate = () => {
+    if (!installed.length) {
+      return null;
+    }
+
+    const list = filterOutRecentWallets([...installed]);
+
+    return list
+      .slice(0, INSTALLED_COUNT)
+      .map(wallet => (
+        <ListWallet
+          key={wallet?.id}
+          imageSrc={AssetUtil.getWalletImage(wallet)}
+          imageHeaders={imageHeaders}
+          name={wallet?.name ?? 'Unknown'}
+          onPress={() => onWalletPress(wallet!)}
+          tagLabel="Installed"
+          tagVariant="success"
+        />
+      ));
   };
 
   const featuredTemplate = () => {
-    if (!featured.length) {
+    if (!featured.length || FEATURED_COUNT < 1) {
       return null;
     }
 
     const list = filterOutRecentWallets([...featured]);
 
     return list
-      .slice(8)
+      .slice(0, FEATURED_COUNT)
       .map(wallet => (
         <ListWallet
           key={wallet?.id}
@@ -56,20 +91,22 @@ export function ConnectView() {
   };
 
   const recommendedTemplate = () => {
-    if (!recommended.length || featured.length) {
+    if (!recommended.length || featured.length || RECOMMENDED_COUNT < 1) {
       return null;
     }
-    const [first, second] = filterOutRecentWallets([...recommended]);
+    const list = filterOutRecentWallets([...recommended]);
 
-    return [first, second].map(wallet => (
-      <ListWallet
-        key={wallet?.id}
-        imageSrc={AssetUtil.getWalletImage(wallet)}
-        imageHeaders={imageHeaders}
-        name={wallet?.name ?? 'Unknown'}
-        onPress={() => onWalletPress(wallet!)}
-      />
-    ));
+    return list
+      .slice(0, RECOMMENDED_COUNT)
+      .map(wallet => (
+        <ListWallet
+          key={wallet?.id}
+          imageSrc={AssetUtil.getWalletImage(wallet)}
+          imageHeaders={imageHeaders}
+          name={wallet?.name ?? 'Unknown'}
+          onPress={() => onWalletPress(wallet!)}
+        />
+      ));
   };
 
   const allWalletsTemplate = () => {
@@ -85,7 +122,7 @@ export function ConnectView() {
   };
 
   const filterOutRecentWallets = (wallets: WcWallet[]) => {
-    const recentIds = recentWallets?.map(wallet => wallet.id);
+    const recentIds = recentWallets?.slice(RECENT_COUNT).map(wallet => wallet.id);
     if (!recentIds?.length) return wallets;
 
     const filtered = wallets.filter(wallet => !recentIds.includes(wallet.id));
@@ -96,6 +133,7 @@ export function ConnectView() {
   return (
     <FlexView padding="s" rowGap="2xs">
       {recentTemplate()}
+      {installedTemplate()}
       {featuredTemplate()}
       {recommendedTemplate()}
       {allWalletsTemplate()}