+++ packages/scaffold/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @web3modal/scaffold-react-native
 
+## 1.0.0-alpha.9
+
+### Patch Changes
+
+- fixes in network view
+- Updated dependencies
+  - @web3modal/core-react-native@1.0.0-alpha.9
+  - @web3modal/ui-react-native@1.0.0-alpha.9
+
 ## 1.0.0-alpha.8
 
 ### Patch Changes